requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
python-dotenv
werkzeug
pymysql  